Generate scaladoc for javalib in Scala 2 (#3035)
* Allow java doc to be published
* Update build
* Add explicit type annotations for unicode char literals in java.lang.Character
* Restore ArrayDequeue javadoc, fix its issues when generating scaladoc
* Enable docs generatin only in javalib for Scala 2


(cherry picked from commit d0b9b3d669a0d43beaa2d59338e767ebffa3fa10)

+++ javalib/src/main/scala/java/lang/process/UnixProcessGen2.scala
@@ -250,7 +250,7 @@ private[lang] class UnixProcessGen2 private (
 
     val fds = stackalloc[struct_pollfd](1.toUInt)
     (fds + 0).fd = pidFd
-    (fds + 0).events = pollEvents.POLLIN | pollEvents.POLLRDNORM
+    (fds + 0).events = (pollEvents.POLLIN | pollEvents.POLLRDNORM).toShort
 
     val tmo = timeout.getOrElse(null)
